refactor: enable strict null checks

+++ src/services/item/plugins/app/appSetting/service.ts
@@ -1,5 +1,6 @@
 import { PermissionLevel } from '@graasp/sdk';
 
+import { MemberCannotAccess } from '../../../../../util/graasp-error';
 import { Repositories } from '../../../../../util/repositories';
 import { validatePermission } from '../../../../authorization';
 import { AppSetting } from './appSettings';
@@ -7,15 +8,18 @@ import { InputAppSetting } from './interfaces/app-setting';
 
 export class AppSettingService {
   async post(
-    memberId: string,
+    memberId: string | undefined,
     repositories: Repositories,
     itemId: string,
     body: Partial<InputAppSetting>,
   ) {
     const { appSettingRepository, memberRepository, itemRepository } = repositories;
+    // check member exists
+    if (!memberId) {
+      throw new MemberCannotAccess();
+    }
     const member = await memberRepository.get(memberId);
 
-    // check item exists? let post fail?
     const item = await itemRepository.get(itemId);
 
     // posting an app setting is allowed to admin only
@@ -25,69 +29,66 @@ export class AppSettingService {
   }
 
   async patch(
-    memberId: string,
+    memberId: string | undefined,
     repositories: Repositories,
     itemId: string,
     appSettingId: string,
     body: Partial<AppSetting>,
   ) {
     const { appSettingRepository, memberRepository, itemRepository } = repositories;
-    // TODO: check member exists
+    // check member exists
+    if (!memberId) {
+      throw new MemberCannotAccess();
+    }
     const member = await memberRepository.get(memberId);
 
     // check item exists? let post fail?
     const item = await itemRepository.get(itemId);
 
     // patching requires admin rights
-    const inheritedMembership = await validatePermission(
-      repositories,
-      PermissionLevel.Admin,
-      member,
-      item,
-    );
+    await validatePermission(repositories, PermissionLevel.Admin, member, item);
 
-    const appSetting = await appSettingRepository.get(appSettingId);
+    await appSettingRepository.get(appSettingId);
 
     return appSettingRepository.patch(itemId, appSettingId, body);
   }
 
   async deleteOne(
-    memberId: string,
+    memberId: string | undefined,
     repositories: Repositories,
     itemId: string,
     appSettingId: string,
   ) {
     const { appSettingRepository, memberRepository, itemRepository } = repositories;
-    // TODO: check member exists
+    // check member exists
+    if (!memberId) {
+      throw new MemberCannotAccess();
+    }
     const member = await memberRepository.get(memberId);
 
     // check item exists? let post fail?
     const item = await itemRepository.get(itemId);
 
     // delete an app data is allowed to admins
-    const inheritedMembership = await validatePermission(
-      repositories,
-      PermissionLevel.Admin,
-      member,
-      item,
-    );
+    await validatePermission(repositories, PermissionLevel.Admin, member, item);
 
     await appSettingRepository.get(appSettingId);
 
     return appSettingRepository.deleteOne(itemId, appSettingId);
   }
 
-  async getForItem(memberId: string, repositories: Repositories, itemId: string) {
+  async getForItem(memberId: string | undefined, repositories: Repositories, itemId: string) {
     const { appSettingRepository, memberRepository, itemRepository } = repositories;
 
-    // TODO: check member exists
-    const member = await memberRepository.get(memberId);
+    // get member if exists
+    // item can be public
+    const member = memberId ? await memberRepository.get(memberId) : undefined;
 
     // check item exists? let post fail?
     const item = await itemRepository.get(itemId);
 
-    // posting an app data is allowed to readers
-    const membership = await validatePermission(repositories, PermissionLevel.Read, member, item);
+    // get app setting is allowed to readers
+    await validatePermission(repositories, PermissionLevel.Read, member, item);
 
     // TODO: get only memberId or with visibility
     return appSettingRepository.getForItem(itemId);
